+++ services/db.php
@@ -30,7 +30,7 @@
 				`id` int(11) NOT NULL  AUTO_INCREMENT PRIMARY KEY,
 				`countryName` varchar(255) NOT NULL,
 				`countryCapitalName` varchar(255) NOT NULL ,
-				`countryPopulation` int(255) NOT NULL )"); 
+				`countryPopulation` DECIMAL(20,3) NOT NULL )"); 
 			}
 		}
 		public function query(string $sql, array $params = []): ?array
